# tests/test_dashboard/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KanbanResponse": {
        "additionalProperties": true,
        "properties": {
          "is_legacy": {
            "title": "Is Legacy",
            "type": "boolean"
          },
          "lanes": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/KanbanTaskData"
              },
              "type": "array"
            },
            "title": "Lanes",
            "type": "object"
          },
          "upgrade_needed": {
            "title": "Upgrade Needed",
            "type": "boolean"
          },
          "weighted_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weighted Percentage"
          }
        },
        "required": [
          "lanes",
          "is_legacy",
          "upgrade_needed"
        ],
        "title": "KanbanResponse",
        "type": "object"
      },
...
      "KanbanTaskData": {
        "additionalProperties": true,
        "properties": {
          "agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent"
          },
          "agent_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Profile"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "encoding_error": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encoding Error"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "lane": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lane"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "prompt_markdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Markdown"
          },
          "prompt_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Path"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "subtasks": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtasks"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "KanbanTaskData",
        "type": "object"
      },
...
      "SyncTriggerFailedResponse": {
        "additionalProperties": true,
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "SyncTriggerFailedResponse",
        "type": "object"
      },
      "SyncTriggerScheduledResponse": {
        "additionalProperties": true,
        "properties": {
          "status": {
            "const": "scheduled",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "SyncTriggerScheduledResponse",
        "type": "object"
      },
      "SyncTriggerSkippedResponse": {
        "additionalProperties": true,
        "properties": {
          "manual_mode": {
            "title": "Manual Mode",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "const": "skipped",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "manual_mode"
        ],
        "title": "SyncTriggerSkippedResponse",
        "type": "object"
      },
      "SyncTriggerUnavailableResponse": {
        "additionalProperties": true,
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "error"
        ],
        "title": "SyncTriggerUnavailableResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP surface for the local Spec Kitty dashboard. All routes are localhost-only. See `kitty-specs/frontend-api-fastapi-openapi-migration-01KQN2JA/` for the contract authoring rules.",
    "title": "Spec Kitty Dashboard API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/kanban/{feature_id}": {
      "get": {
        "operationId": "get_kanban_api_kanban__feature_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "feature_id",
            "required": true,
            "schema": {
              "title": "Feature Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KanbanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Kanban",
        "tags": [
          "kanban"
        ]
      }
    },
...
    "/api/sync/trigger": {
      "post": {
        "operationId": "trigger_sync_api_sync_trigger_post",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SyncTriggerScheduledResponse"
                    },
                    {
                      "$ref": "#/components/schemas/SyncTriggerSkippedResponse"
                    },
                    {
                      "$ref": "#/components/schemas/SyncTriggerUnavailableResponse"
                    },
                    {
                      "$ref": "#/components/schemas/SyncTriggerFailedResponse"
                    }
                  ],
                  "title": "Response Trigger Sync Api Sync Trigger Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Sync",
        "tags": [
          "sync"
        ]
      }
    },
...
  }
}